refactor : compontent로 분리
